chore(docid): doc id in own file
